small fixes after moving to latest duckdb
